+++ app/src/main/java/com/example/android/handystalker/ui/TextRulesActivity.java
@@ -95,10 +95,10 @@ public class TextRulesActivity extends AppCompatActivity {
                 new Handler(Looper.getMainLooper()).post(new Runnable() {
                     @Override
                     public void run() {
-                        if ((int) countPlaces == 0) {
+                        if (countPlaces == 0) {
                             Toast.makeText(getApplicationContext(), getString(R.string.one_place), Toast.LENGTH_LONG).show();
                             return;
-                        } else if ((int) countContacts == 0) {
+                        } else if (countContacts == 0) {
                             Toast.makeText(getApplicationContext(), getString(R.string.one_contact), Toast.LENGTH_LONG).show();
                             return;}
                             else {
